requirements.txt
py2neo~=2021.2.4
icecream~=2.1.3
pyyaml~=6.0.1
python-dotenv~=1.0.1
zhipuai~=2.0.1
loguru~=0.7.2
multipledispatch~=1.0.0
fastapi~=0.109.2
uvicorn~=0.27.1
gradio~=4.19.1
pyahocorasick~=2.0.0
apscheduler~=3.10.4
requests~=2.31.0
edge-tts~=6.1.10
langchain~=0.1.8
langchain-openai~=0.0.7
transformers~=4.38.1
modelscope~=1.12.0
transformers~=4.38.1
pypdf~=4.0.2